third_party/WebKit/LayoutTests/fast/cookies/local-file-can-set-cookies.html

<html>
<head>
<title>Local File Can Set Cookies</title>
<script src="../../resources/js-test.js"></script>
</head>

<body>
<script>
    description('This test checks that a local file can set cookies.  See: rdar://problem/5379090 REGRESSION: Cannot set cookies for local files');

    shouldBe("document.cookie.indexOf('ppkcookie1=testcookie')", "-1");
    document.cookie = 'ppkcookie1=testcookie';
    shouldBeTrue("document.cookie.indexOf('ppkcookie1=testcookie') >= 0");
</script>
</body>
</html>
